requirements.txt
OpenAI==1.102.0
load_dotenv==0.1.0
requests==2.32.5
beautifulsoup4==4.13.5
numpy==2.3.2
pandas==2.3.2
transformers==4.56.0
sentence-transformers== 5.1.0
igraph==0.11.9
